Add files via upload
This a SQL presentation I did with my teammate. To give an overview of what a rational database. I was the team lead on this project since I was the subject expert on SQL DB. I gathered data on my AWS classmates. I used Microsoft Management studio to store my data and performed the demo while presenting in class.